add Quizzes

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -18,7 +18,7 @@ class CreateUsersTable extends Migration
             $table->string('email')->unique();
             $table->timestamp('email_verified_at')->nullable();
             $table->string('password');
-            $table->integer('level')->default(0);
+            $table->tinyInteger('level')->default(0);
             $table->rememberToken();
             $table->timestamps();
         });
